Exception Handling

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -3,11 +3,15 @@ package com.devsuperior.dscommerce.services;
 import com.devsuperior.dscommerce.dto.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DatabaseException;
 import com.devsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -41,13 +45,21 @@ public class ProductService {
     }
 
     @Transactional()
-    public ProductDTO updatet(Long id, ProductDTO dto){
+    public ProductDTO update(Long id, ProductDTO dto){
 
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
+        try {
 
-        return new ProductDTO(entity);
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+
+            return new ProductDTO(entity);
+
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Id não cadastrado em base de dados");
+        }
+
+
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
@@ -59,10 +71,18 @@ public class ProductService {
     }
 
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)//só ira executar o parametro se a chamada da estiver no constexto e outra transação
     public void delete(Long id){
-      repository.deleteById(id);
+        if (!repository.existsById(id)){
+            throw  new ResourceNotFoundException("ID não encontrado");
+        }
 
-
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial pois este produto existe cadastrado em algum pedido");
+        }
     }
+
 }
